Use u32 to store image size
Increase the size of the variables used to store image width and
height from an unsized 8 bit value to an unsized 32 bit value.

+++ src/lib.rs
@@ -9,8 +9,8 @@ pub enum BitDepth {
 }
 
 pub struct Image {
-    pub width:u8,
-    pub height:u8,
+    pub width:u32,
+    pub height:u32,
     pub dat: Vec<u8>,
     pub depth: BitDepth,
 }
